Refactor project structure by merging plugin into main module
Merged the "plugin" directory into the main module to simplify the project structure. Updated file paths accordingly, ensuring all resources and code are moved to the primary module structure for improved maintainability and clarity. Adjusted `settings.gradle.kts` to reflect the changes.
