+++ settings.gradle.kts
@@ -1,3 +1,2 @@
 rootProject.name = "worlds"
 include("api")
-include("plugin")
