+++ src/main/java/net/thenextlvl/worlds/command/WorldCreateCommand.java
@@ -15,7 +15,11 @@ import net.thenextlvl.worlds.WorldsPlugin;
 import net.thenextlvl.worlds.api.model.Generator;
 import net.thenextlvl.worlds.api.model.WorldPreset;
 import net.thenextlvl.worlds.api.preset.Preset;
-import net.thenextlvl.worlds.command.argument.*;
+import net.thenextlvl.worlds.command.argument.DimensionArgument;
+import net.thenextlvl.worlds.command.argument.GeneratorArgument;
+import net.thenextlvl.worlds.command.argument.SeedArgument;
+import net.thenextlvl.worlds.command.argument.WorldPresetArgument;
+import net.thenextlvl.worlds.command.argument.WorldTypeArgument;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
 import org.bukkit.entity.Entity;
