+++ src/main/java/net/thenextlvl/worlds/command/WorldDeleteCommand.java
