+++ src/main/java/net/thenextlvl/worlds/command/WorldLinkRemoveCommand.java
