+++ src/main/java/net/thenextlvl/worlds/command/WorldRegenerateCommand.java
