+++ src/main/java/net/thenextlvl/worlds/command/argument/CommandFlagsArgument.java
