+++ src/main/java/net/thenextlvl/worlds/command/suggestion/GeneratorSuggestionProvider.java
