+++ src/main/java/net/thenextlvl/worlds/model/PaperLevel.java
@@ -8,7 +8,11 @@ import core.nbt.tag.LongTag;
 import lombok.Getter;
 import lombok.experimental.Accessors;
 import net.thenextlvl.worlds.WorldsPlugin;
-import net.thenextlvl.worlds.api.model.*;
+import net.thenextlvl.worlds.api.model.Generator;
+import net.thenextlvl.worlds.api.model.Level;
+import net.thenextlvl.worlds.api.model.LevelBuilder;
+import net.thenextlvl.worlds.api.model.LevelExtras;
+import net.thenextlvl.worlds.api.model.WorldPreset;
 import net.thenextlvl.worlds.api.preset.Preset;
 import org.bukkit.NamespacedKey;
 import org.bukkit.World;
